Move optics annotation to optics package and rename @optic to @optics
