+++ modules/core/arrow-annotations-processor-test/src/main/java/arrow/ap/objects/Optional.kt
@@ -1,8 +1,8 @@
 package arrow.ap.objects
 
-import arrow.OpticsTarget
 import arrow.core.Option
-import arrow.optic
+import arrow.optics.OpticsTarget
+import arrow.optics.optics
 
-@optic([(OpticsTarget.OPTIONAL)])
+@optics([(OpticsTarget.OPTIONAL)])
 data class Optional(val field: String, val nullable: String?, val option: Option<String>)
