+++ modules/core/arrow-annotations-processor-test/src/main/java/arrow/ap/objects/OptionalSealed.kt
@@ -1,10 +1,10 @@
 package arrow.ap.objects
 
-import arrow.OpticsTarget
 import arrow.core.Option
-import arrow.optic
+import arrow.optics.OpticsTarget
+import arrow.optics.optics
 
-@optic([OpticsTarget.OPTIONAL])
+@optics([OpticsTarget.OPTIONAL])
 sealed class OptionalSealed(val field: String, val nullable: String?, val option: Option<String>) {
   data class Optional2(val a: String?) : OptionalSealed("", null, Option.empty())
 }
