+++ modules/core/arrow-annotations-processor/src/main/java/arrow/optics/AnnotationInfo.kt
@@ -1,7 +1,5 @@
 package arrow.optics
 
-import arrow.optic
-
-val opticsAnnotationKClass = optic::class
+val opticsAnnotationKClass = optics::class
 val opticsAnnotationClass = opticsAnnotationKClass.java
 val opticsAnnotationName = "@" + opticsAnnotationKClass.simpleName
