+++ modules/core/arrow-annotations/src/main/java/arrow/optics.kt
@@ -1,11 +1,11 @@
-package arrow
+package arrow.optics
 
 import kotlin.annotation.AnnotationRetention.SOURCE
 import kotlin.annotation.AnnotationTarget.CLASS
 
 @Retention(SOURCE)
 @Target(CLASS)
-annotation class optic(val targets: Array<OpticsTarget> = [OpticsTarget.ISO, OpticsTarget.LENS, OpticsTarget.PRISM, OpticsTarget.OPTIONAL, OpticsTarget.DSL])
+annotation class optics(val targets: Array<OpticsTarget> = [OpticsTarget.ISO, OpticsTarget.LENS, OpticsTarget.PRISM, OpticsTarget.OPTIONAL, OpticsTarget.DSL])
 
 enum class OpticsTarget {
   ISO, LENS, PRISM, OPTIONAL, DSL
